Extract logic further into OfferingsFactory
