+++ common/src/main/java/com/revenuecat/purchases/common/offerings/OfferingsManager.kt
@@ -3,25 +3,18 @@ package com.revenuecat.purchases.common.offerings
 import android.os.Handler
 import android.os.Looper
 import com.revenuecat.purchases.Offerings
-import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.PurchasesError
 import com.revenuecat.purchases.PurchasesErrorCode
 import com.revenuecat.purchases.common.Backend
-import com.revenuecat.purchases.common.BillingAbstract
 import com.revenuecat.purchases.common.LogIntent
-import com.revenuecat.purchases.common.OfferingParser
 import com.revenuecat.purchases.common.caching.DeviceCache
 import com.revenuecat.purchases.common.log
-import com.revenuecat.purchases.models.StoreProduct
 import com.revenuecat.purchases.strings.OfferingStrings
-import org.json.JSONException
-import org.json.JSONObject
 
 class OfferingsManager(
     private val deviceCache: DeviceCache,
     private val backend: Backend,
-    private val billing: BillingAbstract,
-    private val offeringParser: OfferingParser,
+    private val offeringsFactory: OfferingsFactory,
     // This is nullable due to: https://github.com/RevenueCat/purchases-flutter/issues/408
     private val mainHandler: Handler? = Handler(Looper.getMainLooper())
 ) {
@@ -71,73 +64,26 @@ class OfferingsManager(
             appUserID,
             appInBackground,
             { offeringsJSON ->
-                try {
-                    val allRequestedProductIdentifiers = extractProductIdentifiers(offeringsJSON)
-                    if (allRequestedProductIdentifiers.isEmpty()) {
-                        handleErrorFetchingOfferings(
-                            PurchasesError(
-                                PurchasesErrorCode.ConfigurationError,
-                                OfferingStrings.CONFIGURATION_ERROR_NO_PRODUCTS_FOR_OFFERINGS
-                            ),
-                            onError
-                        )
-                    } else {
-                        getStoreProductsById(allRequestedProductIdentifiers, { productsById ->
-                            logMissingProducts(allRequestedProductIdentifiers, productsById)
-
-                            val offerings = offeringParser.createOfferings(offeringsJSON, productsById)
-                            if (offerings.all.isEmpty()) {
-                                handleErrorFetchingOfferings(
-                                    PurchasesError(
-                                        PurchasesErrorCode.ConfigurationError,
-                                        OfferingStrings.CONFIGURATION_ERROR_PRODUCTS_NOT_FOUND
-                                    ),
-                                    onError
-                                )
-                            } else {
-                                synchronized(this@OfferingsManager) {
-                                    deviceCache.cacheOfferings(offerings)
-                                }
-                                dispatch {
-                                    onSuccess?.invoke(offerings)
-                                }
-                            }
-                        }, { error ->
-                            handleErrorFetchingOfferings(error, onError)
-                        })
+                offeringsFactory.createOfferings(
+                    offeringsJSON,
+                    onError = { error ->
+                        handleErrorFetchingOfferings(error, onError)
+                    },
+                    onSuccess = { offerings ->
+                        synchronized(this@OfferingsManager) {
+                            deviceCache.cacheOfferings(offerings)
+                        }
+                        dispatch {
+                            onSuccess?.invoke(offerings)
+                        }
                     }
-                } catch (error: JSONException) {
-                    log(LogIntent.RC_ERROR, OfferingStrings.JSON_EXCEPTION_ERROR.format(error.localizedMessage))
-                    handleErrorFetchingOfferings(
-                        PurchasesError(
-                            PurchasesErrorCode.UnexpectedBackendResponseError,
-                            error.localizedMessage
-                        ),
-                        onError
-                    )
-                }
+                )
             }, { error ->
                 handleErrorFetchingOfferings(error, onError)
             })
     }
 
     private fun isOfferingsCacheStale(appInBackground: Boolean) = deviceCache.isOfferingsCacheStale(appInBackground)
-
-    private fun extractProductIdentifiers(offeringsJSON: JSONObject): Set<String> {
-        val jsonOfferingsArray = offeringsJSON.getJSONArray("offerings")
-        val productIds = mutableSetOf<String>()
-        for (i in 0 until jsonOfferingsArray.length()) {
-            val jsonPackagesArray =
-                jsonOfferingsArray.getJSONObject(i).getJSONArray("packages")
-            for (j in 0 until jsonPackagesArray.length()) {
-                jsonPackagesArray.getJSONObject(j)
-                    .optString("platform_product_identifier").takeIf { it.isNotBlank() }?.let {
-                        productIds.add(it)
-                    }
-            }
-        }
-        return productIds
-    }
 
     private fun handleErrorFetchingOfferings(
         error: PurchasesError,
@@ -159,53 +105,6 @@ class OfferingsManager(
             onError?.invoke(error)
         }
     }
-
-    private fun getStoreProductsById(
-        productIds: Set<String>,
-        onCompleted: (Map<String, List<StoreProduct>>) -> Unit,
-        onError: (PurchasesError) -> Unit
-    ) {
-        billing.queryProductDetailsAsync(
-            ProductType.SUBS,
-            productIds,
-            { subscriptionProducts ->
-                val productsById = subscriptionProducts
-                    .groupBy { subProduct -> subProduct.purchasingData.productId }
-                    .toMutableMap()
-                val subscriptionIds = productsById.keys
-
-                val inAppProductIds = productIds - subscriptionIds
-                if (inAppProductIds.isNotEmpty()) {
-                    billing.queryProductDetailsAsync(
-                        ProductType.INAPP,
-                        inAppProductIds,
-                        { inAppProducts ->
-                            productsById.putAll(inAppProducts.map { it.purchasingData.productId to listOf(it) })
-                            onCompleted(productsById)
-                        }, {
-                            onError(it)
-                        }
-                    )
-                } else {
-                    onCompleted(productsById)
-                }
-            }, {
-                onError(it)
-            })
-    }
-
-    private fun logMissingProducts(
-        allProductIdsInOfferings: Set<String>,
-        storeProductByID: Map<String, List<StoreProduct>>
-    ) = allProductIdsInOfferings
-        .filterNot { storeProductByID.containsKey(it) }
-        .takeIf { it.isNotEmpty() }
-        ?.let { missingProducts ->
-            log(
-                LogIntent.GOOGLE_WARNING, OfferingStrings.CANNOT_FIND_PRODUCT_CONFIGURATION_ERROR
-                    .format(missingProducts.joinToString(", "))
-            )
-        }
 
     private fun dispatch(action: () -> Unit) {
         if (Thread.currentThread() != Looper.getMainLooper().thread) {
