+++ purchases/src/main/kotlin/com/revenuecat/purchases/PurchasesFactory.kt
@@ -21,6 +21,7 @@ import com.revenuecat.purchases.common.diagnostics.DiagnosticsFileHelper
 import com.revenuecat.purchases.common.diagnostics.DiagnosticsSynchronizer
 import com.revenuecat.purchases.common.diagnostics.DiagnosticsTracker
 import com.revenuecat.purchases.common.networking.ETagManager
+import com.revenuecat.purchases.common.offerings.OfferingsFactory
 import com.revenuecat.purchases.common.offerings.OfferingsManager
 import com.revenuecat.purchases.common.offlineentitlements.OfflineCustomerInfoCalculator
 import com.revenuecat.purchases.common.offlineentitlements.OfflineEntitlementsManager
@@ -175,8 +176,7 @@ internal class PurchasesFactory(
             val offeringsManager = OfferingsManager(
                 cache,
                 backend,
-                billing,
-                offeringParser
+                OfferingsFactory(billing, offeringParser)
             )
 
             return Purchases(
